+++ DataAccessObjects/DAO/ReportDAO.cs
@@ -1,4 +1,5 @@
 ﻿using BusinessObjects;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -15,33 +16,64 @@ namespace DataAccessObjects.DAO
         {
             _context = context;
         }
-        
-        public Report GetReport(int id)
+
+        private static ReportDAO instance = null;
+        public static readonly object Lock = new object();
+        private ReportDAO() { }
+        public static ReportDAO Instance
         {
-            Report report = _context.Reports.Find(id);
-            return report;
+            get
+            {
+                lock (Lock)
+                {
+                    if (instance == null)
+                    {
+                        instance = new ReportDAO();
+                    }
+                    return instance;
+                }
+            }
         }
 
-        public List<Report> GetReports()
+        public async Task<List<Report>> GetReports()
         {
-            List<Report> reports = _context.Reports.ToList();
-            return reports;
-        }
-        public void CreateReport(Report report)
-        {
-             _context.Reports.Add(report);
+            return await _context.Reports.ToListAsync();
         }
 
-        public void UpdateReport(Report report)
+        public async Task<Report> GetReportById(int id)
         {
-            Report currentReport = _context.Reports.Find(report.Id);
-            _context.Reports.Update(currentReport);
+            return await _context.Reports.SingleOrDefaultAsync(r => r.Id == id);
         }
 
-        public void DeleteReport(int id)
+        public async Task AddReport(Report report)
         {
-            Report report = _context.Reports.Find(id);
-            _context.Reports.Remove(report);
+            await _context.Reports.AddAsync(report);
+            await _context.SaveChangesAsync();
+        }
+
+        public async Task UpdateReport(Report report)
+        {
+            var existingReport = await _context.Reports.FindAsync(report.Id);
+            if (existingReport != null)
+            {
+                existingReport.ReporterId = report.ReporterId;
+                existingReport.ProductId = report.ProductId;
+                existingReport.Status = report.Status;
+                existingReport.UpdatedAt = DateTime.Now;
+
+                _context.Reports.Update(existingReport);
+                await _context.SaveChangesAsync();
+            }
+        }
+
+        public async Task RemoveReport(int id)
+        {
+            var report = await _context.Reports.FindAsync(id);
+            if (report != null)
+            {
+                _context.Reports.Remove(report);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
